MVP implementation of injecting JavaScript in to head or foot section
Add ability to specify text (ideally HTML but no validation around this)
for "Head" or "Foot" that will be added to the "HeadScripts" &
"FootScripts" via the resource manager on every page.

This can be managed via the admin dashboard in a sub menu option under
the configuration admin menu option.